src/main/java/com/example/urlparser/service/UrlCollectorService.java
package com.example.urlparser.service;

import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;
import org.springframework.web.client.RestTemplate;

@Service
public class UrlCollectorService {

    /*
    TODO url validation
     */
    public String getContent(String url) {
        RestTemplate restTemplate = new RestTemplate();
        ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
        return response.getBody();
    }
}
